ufmf: return image as uint8
git-svn-id: https://code.astraw.com/kookaburra/trunk/ufmf@3436 e6c0db7a-28da-0310-a2fc-f6ab65d9114a

+++ motmot/ufmf/ufmf.py
@@ -283,7 +283,7 @@ class FlyMovieEmulator(object):
                     more['sumsqf'] = sumsqf_image
                 else:
                     mean_image = tmp
-                self._last_frame = np.array(mean_image,copy=True)
+                self._last_frame = np.array(mean_image,copy=True).astype(np.uint8)
                 more['mean'] = mean_image
             elif self.white_background:
                 self._last_frame = numpy.empty(self._bg0.shape,dtype=np.uint8)
